Fixed64 namespace

+++ include/FixedMath.h
@@ -149,7 +149,7 @@ namespace Fixed64
       return t3;
     }
 
-    static Fixed64 Rcp(Fixed64 num) { return Fixed64(4294967296LL / num.value); }
+    static Fixed64 Rcp(Fixed64 num) { return Fixed64(static_cast<int64_t>(4294967296LL / num.value)); }
 
     static Fixed64 Rsqrt(Fixed64 num)
     {
